Seat 예약 방법 수정

+++ Libaray/src/main/java/com/Libaray/app/domain/SeatDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class SeatDTO {
 
-	private int connectionseq;
+	private Long connectionseq;
 	private int seatId;
 	private String seatName;
 	private int studentNumber;
